Added some fixes

+++ rateService/src/main/java/rateservice/entity/ClientFeedback.java
@@ -17,7 +17,7 @@ public class ClientFeedback {
     private Long id;
 
     @Column(name = "rate",nullable = false)
-    private Double rate;
+    private Long rate;
 
     @Column(nullable = true)
     private String comment;
